refactor(analytics): group telemetry components under components/telemetry/
These are render-null effect components, not UI. Keeping them alongside
real UI components made the components/ folder noisy. A dedicated
subfolder makes their semantic role obvious at a glance.
