+++ src/components/telemetry/CompareTelemetry.tsx
