+++ src/components/telemetry/LensDetailTelemetry.tsx
